refactor(descriptor): 네이밍 변경
