+++ packages/core/src/descriptors/schema.ts
